feat/upd getDeadlineAndBirthDate - create getCatAgeInMonth method and upd date, use age from getDeadlineAndBirthDate in CatInfoItem
